refactor(core): move autostart logic out of sysopt and clean up naming
