+++ src-tauri/src/core/sysopt.rs
@@ -1,13 +1,8 @@
-#[cfg(target_os = "windows")]
-use crate::utils::schtasks as startup_task;
 use crate::{
     config::{Config, IVerge},
-    core::handle::Handle,
     singleton,
 };
 use anyhow::Result;
-#[cfg(not(target_os = "windows"))]
-use clash_verge_logging::logging_error;
 use clash_verge_logging::{Type, logging};
 use parking_lot::RwLock;
 use scopeguard::defer;
@@ -20,10 +15,6 @@ use std::{
     time::Duration,
 };
 use sysproxy::{Autoproxy, GuardMonitor, GuardType, Sysproxy};
-#[cfg(not(target_os = "windows"))]
-use tauri_plugin_autostart::ManagerExt as _;
-#[cfg(target_os = "windows")]
-use tauri_plugin_clash_verge_sysinfo::is_current_app_handle_admin;
 
 pub struct Sysopt {
     update_sysproxy: AtomicBool,
@@ -227,65 +218,5 @@ impl Sysopt {
         auto.set_auto_proxy()?;
 
         Ok(())
-    }
-
-    /// update the startup
-    pub async fn update_launch(&self) -> Result<()> {
-        let enable_auto_launch = { Config::verge().await.latest_arc().enable_auto_launch };
-        let is_enable = enable_auto_launch.unwrap_or(false);
-        logging!(info, Type::System, "Setting auto-launch state to: {:?}", is_enable);
-
-        #[cfg(target_os = "windows")]
-        {
-            let is_admin = is_current_app_handle_admin(Handle::app_handle());
-            startup_task::set_auto_launch(is_enable, is_admin).await
-        }
-
-        #[cfg(not(target_os = "windows"))]
-        {
-            self.try_original_autostart_method(is_enable);
-            Ok(())
-        }
-    }
-
-    /// 尝试使用原来的自启动方法
-    #[cfg(not(target_os = "windows"))]
-    fn try_original_autostart_method(&self, is_enable: bool) {
-        let app_handle = Handle::app_handle();
-        let autostart_manager = app_handle.autolaunch();
-
-        if is_enable {
-            logging_error!(Type::System, "{:?}", autostart_manager.enable());
-        } else {
-            logging_error!(Type::System, "{:?}", autostart_manager.disable());
-        }
-    }
-
-    /// 获取当前自启动的实际状态
-    pub fn get_launch_status(&self) -> Result<bool> {
-        #[cfg(target_os = "windows")]
-        {
-            let enabled = startup_task::is_auto_launch_enabled();
-            if let Ok(status) = enabled {
-                logging!(info, Type::System, "Auto launch status (scheduled task): {status}");
-            }
-            enabled
-        }
-
-        #[cfg(not(target_os = "windows"))]
-        {
-            let app_handle = Handle::app_handle();
-            let autostart_manager = app_handle.autolaunch();
-            match autostart_manager.is_enabled() {
-                Ok(status) => {
-                    logging!(info, Type::System, "Auto launch status: {status}");
-                    Ok(status)
-                }
-                Err(e) => {
-                    logging!(error, Type::System, "Failed to get auto launch status: {e}");
-                    Err(anyhow::anyhow!("Failed to get auto launch status: {}", e))
-                }
-            }
-        }
     }
 }
